refactor some of the code to separate files and begin a generic message interface
