+++ .vscode/settings.json
@@ -5,5 +5,7 @@
             "source.fixAll": "always"
         },
         "editor.defaultFormatter": "charliermarsh.ruff"
-    }
+    },
+    "python.analysis.typeCheckingMode": "standard"
+    
 }
