+++ lib/modelConfirmationGUI.py
@@ -1,0 +1,51 @@
+from .modelHelpers import GPTState
+from talon import imgui, actions, Module, clip
+
+
+mod = Module()
+
+
+@imgui.open()
+def confirmation_gui(gui: imgui.GUI):
+    gui.text("Confirm model output before pasting")
+    gui.line()
+    gui.spacer()
+    gui.text(GPTState.text_to_confirm)
+
+    gui.spacer()
+    if gui.button("Paste model output"):
+        actions.user.paste_model_confirmation_gui()
+
+    gui.spacer()
+    if gui.button("Copy model output"):
+        actions.user.copy_model_confirmation_gui()
+
+    gui.spacer()
+    if gui.button("Deny model output"):
+        actions.user.close_model_confirmation_gui()
+
+
+@mod.action_class
+class UserActions:
+    def add_to_confirmation_gui(model_output: str):
+        """Add text to the confirmation gui"""
+        GPTState.text_to_confirm = model_output
+        confirmation_gui.show()
+
+    def close_model_confirmation_gui():
+        """Close the model output without pasting it"""
+        GPTState.text_to_confirm = ""
+        confirmation_gui.hide()
+
+    def copy_model_confirmation_gui():
+        """Copy the model output to the clipboard"""
+        clip.set_text(GPTState.text_to_confirm)
+        GPTState.text_to_confirm = ""
+
+        confirmation_gui.hide()
+
+    def paste_model_confirmation_gui():
+        """Paste the model output"""
+        actions.user.paste(GPTState.text_to_confirm)
+        GPTState.text_to_confirm = ""
+        confirmation_gui.hide()
